pyproject.toml
[project]
name = "xployt_lvl2"
version = "0.1.0"
description = ""
authors = [
    {name = "Savith Panamgama",email = "geethikasavith@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "openai (>=1.93.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "pystache (>=0.6.8,<0.7.0)",
    "dir-tree (>=0.0.2,<0.0.3)",
    "rich (>=14.0.0,<15.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "uvicorn main:app --reload"
